pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opengrasp"
version = "0.1.2"
description = "Open-source autonomous job search agent. Runs locally."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = [
  "job-search",
  "cli",
  "ollama",
  "local-first",
  "career",
  "resume",
  "ats",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Office/Business",
]
dependencies = [
  "typer>=0.12",
  "rich>=13",
  "textual>=0.80",
  "ollama>=0.3",
  "playwright>=1.44",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "pyyaml>=6",
  "jinja2>=3.1",
  "httpx>=0.27",
  "python-slugify",
  "python-dateutil",
]

[project.scripts]
opengrasp = "cli.main:app"

[tool.setuptools.package-data]
agent = ["prompts/*.md", "templates/*.html"]
cli = ["assets/*.yml"]

[tool.setuptools.packages.find]
include = ["agent", "agent.*", "cli", "cli.*", "memory", "memory.*"]
